* Only load assets when needed and don't add version to .js files
This commit splits the admin.js file and implements a system to only load the necessary css/js for some components. This prevents conflicts that occurr when an asset is loaded twice by two different plugins. For this same reason, this commit no longer adds the lib version to javascript handles. If a script is requested twice, it should only be enqueued once or we can run into scripting errors.